--- a/package.json
+++ b/package.json
@@ -40,19 +40,19 @@
     "@types/long": "^4.0.0",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.7.0",
-    "chai": "^4.1.0",
+    "chai": "^4.2.0",
     "codecov": "^3.0.4",
     "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
+    "nyc": "^13.1.0",
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
-    "tslint-config-standard": "^7.1.0",
+    "tslint-config-standard": "^8.0.1",
     "typescript": "^3.0.1"
   },
   "dependencies": {
     "bignumber.js": "^7.2.1",
     "extensible-error": "^1.0.2",
     "long": "^4.0.0",
-    "oer-utils": "^3.0.1"
+    "oer-utils": "^3.1.1"
   }
 }