--- a/package.json
+++ b/package.json
@@ -50,7 +50,6 @@
     "typescript": "^3.0.1"
   },
   "dependencies": {
-    "bignumber.js": "^7.2.1",
     "extensible-error": "^1.0.2",
     "long": "^4.0.0",
     "oer-utils": "^3.1.1"