Updated to latest `oer-utils` with improved performance.

Did a quick [benchmark of bignumber.js vs long](https://gist.github.com/justmoon/90300c1e264e99ecefcdd7f2014816fb). Long is ~33 times faster:

```
Long x 5,546 ops/sec ±2.75% (84 runs sampled)
Bignumber.js x 165 ops/sec ±1.94% (74 runs sampled)
Fastest is Long
```

This also allows us removing `bignumber.js` as a dependency of `ilp-packet`. Obviously, it' still an indirect dependency via `oer-utils` but it isn't used at all for (de)serializing ilp-packets. It might be worth making an `oer-utils-light` which doesn't depend on bignumber.js at all. I don't think anything in ILP actually needs support for >64 bit numbers.

Edit: Updated the benchmark to add deserialization. About 10x difference.

```
Long serialize x 8,195 ops/sec ±3.41% (81 runs sampled)
Bignumber.js serialize x 136 ops/sec ±9.00% (63 runs sampled)
Long deserialize x 7,317 ops/sec ±2.83% (81 runs sampled)
Bignumber.js deserialize x 708 ops/sec ±1.95% (83 runs sampled)
```