@sharafian 2018-10-23
Even for those totals we could theoretically wrap them in a class that uses long until it exceeds 64 bits and then switched to bignum