@justmoon 2018-10-24
I think it would be totally reasonable to cap our STREAM reference implementation at UInt64. For reference, that would introduce a limit of $18.4 billion per STREAM connection at nanodollar resolution. So I wouldn't edit the spec to say that STREAM - the protocol - has a cap of UInt64 but we may want to add a note that *some implementations* may have a cap.

If you are actually going to send that kind of money over STREAM, I would hope you'd be using a more enterprise-y implementation than the JavaScript reference one anyway.

cc/ @sentientwaffle @emschwartz